Changed floats to ints
Because processing uses pixels as int coordinates, exact coordinates using ints works better

+++ Point.pde
@@ -1,6 +1,6 @@
 class Point{
-  float x,y;
-  Point(float a, float b){
+  int x,y;
+  Point(int a, int b){
     x = a;
     y = b;
   }
